minor finishing touches
